Split up test to multiple files
