+++ package.json
@@ -42,13 +42,16 @@
     "npm": "^2.0.0 || ^3.0.0 || ^4.0.0"
   },
   "scripts": {
+    "install": "node-gyp rebuild",
     "postinstall": "node-gyp rebuild",
     "test": "npm run --silent mocha && npm run --silent lint",
-    "mocha": "mocha-parallel-tests --slow 500 --timeout 10000 test/",
+    "test:run": "npm run --silent mocha:run --",
+    "mocha": "npm run --silent mocha:run:all",
+    "mocha:run": "mocha-parallel-tests --slow 500 --timeout 10000",
+    "mocha:run:all": "npm run --silent mocha:run -- test/",
     "lint": "npm run --silent eslint --",
     "eslint": "eslint ./src ./test ./samples getdns.js",
-    "eslint:fix": "eslint --fix ./src ./test ./samples getdns.js",
-    "install": "node-gyp rebuild"
+    "eslint:fix": "eslint --fix ./src ./test ./samples getdns.js"
   },
   "license": "BSD-3-Clause",
   "bugs": {
